Data Preparation : No Feature Selection will be Done
All the features we have identify the face. The features are the RGB values of every pixel of the image cropped around the face. So, in order not to lose any accuracy on the model or misidentify some pictures, we will be keeping all the features to be used with CNN.